work in progress

+++ src/main/scala/denali/tasks/Task.scala
@@ -13,7 +13,7 @@ sealed trait Task {
 
 case class InitialSearchTask(globalOptions: GlobalOptions,
                              instruction: Instruction,
-                             budget: Int) extends Task {
+                             budget: Long) extends Task {
   override def toString = {
     s"initial search for $instruction with budget=$budget"
   }
